- remove unnecessary checks of whether '_token' is in request.POST because util function get_user does that - remove duplicate change_password and send_verification API implementation (one in views_user, one in views_event)

+++ mocacares_backend/event_platform/views/views_event.py
@@ -19,9 +19,6 @@ def get_comments(request):
 
 
 def get_my_comments(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
     user = get_user(request)
     if isinstance(user, AnonymousUser):
         return response_of_failure(msg='You need to log in to view your comments.')
@@ -31,9 +28,6 @@ def get_my_comments(request):
 
 
 def update_system_config(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
     user = get_user(request)
     if isinstance(user, AnonymousUser):
         return response_of_failure(msg='You need to log in to update your preferences.')
@@ -84,9 +78,6 @@ def get_events(request):
 
 
 def add_event(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
     user = get_user(request)
     if isinstance(user, AnonymousUser):
         return response_of_failure(msg='You need to log in to post an event.')
@@ -137,20 +128,17 @@ def add_event(request):
 
 def delete_event(request):
     user = get_user(request)
+    if isinstance(user, AnonymousUser):
+        return response_of_failure(msg='You need to log in first.')
+
     user_type = user.user_type
     if user_type != 2:
-        return JsonResponse({
-            'code': 0,
-            'msg': 'No permission'
-        })
+        return response_of_failure(msg='No permission')
     event_id = request.POST['aid']
     try:
         event = Event.objects.get(pk=event_id)
         event.delete()
-        return JsonResponse({
-            'code': 1,
-            'msg': 'Success'
-        })
+        return response_of_success(msg='Success')
     except ObjectDoesNotExist:
         return response_of_failure(msg='event cannot be found')
 
@@ -173,71 +161,7 @@ def get_event_types(request):
     return JsonResponse(api_returned_object(info=list(event_types)), encoder=EventTypeEncoder)
 
 
-def send_verify(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
-    user = get_user(request)
-    if isinstance(user, AnonymousUser):
-        return response_of_failure(msg='You need to log in to change password.')
-
-    token = request.POST['_token']
-
-    verification_code = random.randint(0, 99999)
-    verification_code_str = str(verification_code).zfill(5)
-
-    save_token_vericode_pair(token=token, verification_code=verification_code_str)
-
-    EmailThread(
-        subject = 'MocaCare Verification Code',
-        content = 'Your verification code is {}'.format(verification_code_str),
-        receiver_list = [user.email_address]
-    ).start()
-
-    #TODO: Image OneToOne
-
-    return JsonResponse({
-        'code': 1,
-        'msg': 'Verification code sent successfully.'
-    }, status=200)
-
-
-def change_pwd(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
-    user = get_user(request)
-    if isinstance(user, AnonymousUser):
-        return response_of_failure(msg='You need to log in to change password.')
-
-    if 'verify' not in request.POST:
-        return response_of_failure(msg='Verification code not provided.')
-
-    if 'newpwd' not in request.POST:
-        return response_of_failure(msg='New password not provided.')
-
-    verification_code = request.POST['verify']
-    new_pwd = request.POST['newpwd']
-    token = request.POST['_token']
-
-    if not token_vericode_pair_exists(token=token, verification_code=verification_code):
-        return response_of_failure(msg='Invalid token or verification code.')
-
-    if not validate_password_format(new_pwd):
-        return response_of_failure(msg='password must contain at least 6 digits')
-    user.set_password(new_pwd)
-    user.save()
-
-    return JsonResponse({
-        'code': 1,
-        'msg': 'Password changed successfully.'
-    }, status=200)
-
-
 def get_recommended_events(request):
-    if '_token' not in request.POST:
-        return response_of_failure(msg='Invalid token')
-
     user = get_user(request)
     if isinstance(user, AnonymousUser):
         return response_of_failure(msg='You need to log in to view events.')
